requirements.txt
click==8.0.3
Flask==0.11
flask-mongoengine==0.8.2
Flask-Script==2.0.5
Flask-WTF==1.0.0
importlib-metadata==4.8.2
injector==0.18.4
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==2.5.1
MarkupSafe==2.0.1
mongoengine==0.11.0
py-bcrypt==0.4
pymongo==3.4.0
python-dotenv==0.19.2
six==1.16.0
typing-extensions==4.0.1
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
